Move code from rustc_trait_selection/opaque_types to better places
